fix: describe remaining retry eligibility without promising recovery
